Refactor: Split test_preview_sync out from test_preview.
